Made a telegram2discord message converter to dry out the code a bit.
