+++ lib/telegram2discord/setup.js
@@ -7,66 +7,35 @@
 const Application = require("../Application");
 const updateGetter = require("./updategetter");
 const handleEntities = require("./handleEntities");
+const messageConverter = require("./messageConverter");
 const MessageMap = require("../MessageMap");
 const _ = require("lodash");
-
-/********************
- * Helper functions *
- ********************/
-
-/**
- * Gets the display name of a user
- *
- * @param {Object} user	A user object
- * @param {Object} chat	A chat object
- *
- * @return {String}	The user's display name
- *
- * @private
- */
-function getDisplayName(user, chat) {
-	// Handle for the result
-	let displayName = null;
-
-	try {
-		// Extract the username
-		displayName = user.username;
-
-		// Check whether or not to use names instead (or if the username does not exist
-		if (!displayName || Application.settings.telegram.useFirstNameInsteadOfUsername) {
-			displayName = user.first_name;
-		}
-	} catch (err) {
-		// No user given. Use the chat title instead
-		displayName = chat.title;
-	}
-
-	return displayName;
-}
 
 /**
  * Creates a function which sends files from Telegram to discord
  *
  * @param {BotAPI} tgBot	The Telegram bot
  * @param {Discord.Client} dcBot	The Discord bot
+ * @param {DiscordUserMap} arg.dcUsers	A map between discord users and their IDs
  *
  * @returns {Function}	A function which can be used to send files from Telegram to Discord
  *
  * @private
  */
-function makeFileSender(tgBot, dcBot) {
+function makeFileSender(tgBot, dcBot, dcUsers) {
 	/**
 	 * Sends a file to Discord
 	 *
-	 * @param {String} arg.fromName	Display name of the sender
+	 * @param {Message} arg.message	Display name of the sender
 	 * @param {String} arg.fileId	ID of the file to download from Telegram's servers
 	 * @param {String} arg.fileName	Name of the file to send
 	 * @param {String} [arg.caption]	Additional text to send with the file
 	 * @param {Boolean} [arg.resolveExtension]	Set to true if the bot should try to find the file extension itself, in which case it will be appended to the file name. Defaults to false
 	 */
-	return function({fromName, fileId, fileName, caption = "", resolveExtension = false}) {
+	return function({message, fileId, fileName, caption = "", resolveExtension = false}) {
 		// Make the text to send
-		let textToSend = `**${fromName}**:\n${caption}`
+		let messageObj = messageConverter({ message, dcUsers, tgBot });
+		let textToSend = `**${messageObj.from}**:\n${caption}`
 
 		// Handle for the file extension
 		let extension = "";
@@ -163,7 +132,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	updateGetter(tgBot, Application.settings);
 
 	// Make the file sender
-	const sendFile = makeFileSender(tgBot, dcBot);
+	const sendFile = makeFileSender(tgBot, dcBot, dcUsers);
 
 	// Create the message handler wrapper
 	const wrapFunction = createMessageHandler(tgBot);
@@ -171,41 +140,11 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for text messages from Telegram
 	tgBot.on("text", wrapFunction((message) => {
 
-		// Convert the text to Discord format
-		let text = handleEntities(dcUsers, message.text, message.entities);
-
-		// Find out who the message is from
-		let fromName = getDisplayName(message.from, message.chat);
-
-		// Check if it is a reply
-		if (message.reply_to_message !== undefined) {
-			// Get the name of the user this is a reply to
-			let inReplyTo = getDisplayName(message.reply_to_message.from, message.chat);
-
-			// Is this a reply to the bot, i.e. to a Discord user?
-			if (message.reply_to_message.from.id === tgBot.me.id) {
-				// Get the name of the Discord user this is a reply to
-				// TODO Maybe a mapping, so the name doesn't have to be parsed out of the message?
-				let dcUsername = message.reply_to_message.text.split("\n")[0];
-				inReplyTo = dcUsers.lookupUsername(dcUsername) ? `<@${dcUsers.lookupUsername(dcUsername)}>` : dcUsername;
-			}
-
-			// Add it to the 'from' text
-			fromName = `${fromName} (in reply to ${inReplyTo})`;
-		}
-
-		// Check if it is a forwarded message
-		if (message.forward_from !== undefined) {
-			// Find the name of the user this was forwarded from
-			let forwardFrom = getDisplayName(message.forward_from, message.chat);
-
-			// Add it to the 'from' text
-			fromName = `${forwardFrom} (forwarded by ${fromName})`;
-		}
+		let messageObj = messageConverter({ message, dcUsers, tgBot });
 
 		// Pass it on to Discord when the dcBot is ready
 		dcBot.ready.then(() => {
-			return dcBot.channels.get(Application.settings.discord.channelID).send(`**${fromName}**: ${text}`);
+			return dcBot.channels.get(Application.settings.discord.channelID).send(messageObj.composed);
 		  })
 		  .then((dcMessage) => {
 			// Make the mapping so future edits can work
@@ -220,7 +159,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for photo messages from Telegram
 	tgBot.on("photo", wrapFunction((message) => {
 		sendFile({
-			fromName: getDisplayName(message.from, message.chat),
+			message,
 			fileId: message.photo[message.photo.length-1].file_id,
 			fileName: "photo.jpg",	// Telegram will convert it to jpg no matter what filetype is actually sent
 			caption: message.caption
@@ -230,7 +169,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for stickers from Telegram
 	tgBot.on("sticker", wrapFunction((message) => {
 		sendFile({
-			fromName: getDisplayName(message.from, message.chat),
+			message,
 			fileId: message.sticker.thumb.file_id,
 			fileName: "sticker.webp",	// Telegram will insist that it is a jpg, but it really is a webp
 			caption: message.sticker.emoji
@@ -240,7 +179,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for filetypes not caught by the other filetype handlers
 	tgBot.on("document", (message) => {
 		sendFile({
-			fromName: getDisplayName(message.from, message.chat),
+			message,
 			fileId: message.document.file_id,
 			fileName: message.document.file_name
 		});
@@ -249,7 +188,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for audio messages
 	tgBot.on("audio", (message) => {
 		sendFile({
-			fromName: getDisplayName(message.from, message.chat),
+			message,
 			fileId: message.audio.file_id,
 			fileName: message.audio.title,
 			resolveExtension: true
@@ -259,7 +198,7 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 	// Set up event listener for video messages
 	tgBot.on("video", (message) => {
 		sendFile({
-			fromName: getDisplayName(message.from, message.chat),
+			message,
 			fileId: message.video.file_id,
 			fileName: "video",
 			resolveExtension: true
@@ -272,33 +211,16 @@ function setup(tgBot, dcBot, dcUsers, messageMap) {
 		dcBot.ready.then(() => messageMap.getCorresponding(MessageMap.TELEGRAM_TO_DISCORD, tgMessage.message_id))	// Try to get the corresponding message in Discord
 		  .then((dcMessageId) => dcBot.channels.get(Application.settings.discord.channelID).fetchMessage(dcMessageId))	// Get the message from Discord
 		  .then((dcMessage) => {
-			// Convert the text to Discord format
-			tgMessage.text = handleEntities(dcUsers, tgMessage.text, tgMessage.entities);
 
-			// Find out who the message is from
-			let fromName = getDisplayName(tgMessage.from, tgMessage.chat);
-
-			// Check if it is a reply
-			if (tgMessage.reply_to_message !== undefined) {
-				// Get the name of the user this is a reply to
-				let inReplyTo = getDisplayName(tgMessage.reply_to_message.from, tgMessage.chat);
-
-				// Is this a reply to the bot, i.e. to a Discord user?
-				if (tgMessage.reply_to_message.from.id === tgBot.me.id) {
-					// Get the name of the Discord user this is a reply to
-					// TODO Maybe a mapping, so the name doesn't have to be parsed out of the message?
-					let dcUsername = tgMessage.reply_to_message.text.split("\n")[0];
-					inReplyTo = dcUsers.lookupUsername(dcUsername) ? `<@${dcUsers.lookupUsername(dcUsername)}>` : dcUsername;
-				}
-
-				// Add it to the 'from' text
-				fromName = `${fromName} (in reply to ${inReplyTo})`;
-			}
+			let messageObj = messageConverter({ message: tgMessage, dcUsers, tgBot });
 
 			// Try to edit the message
-			return dcMessage.edit(`**${fromName}**: ${tgMessage.text}`);
+			return dcMessage.edit(messageObj.composed);
 		  })
-		  .catch((err) => Application.logger.error("Could not edit Discord message:", err));
+		  .catch((err) => {
+			Application.logger.error("Could not edit Discord message:", err);
+			Application.logger.error("Failed message:", message.text);
+		});
 	});
 
 	// Make a promise which resolves when the dcBot is ready
